SEL4C/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
idna==3.4
mariadb==1.1.7
mysqlclient==2.2.0
packaging==23.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.5
